fix: 修复新创建Plan时无法创建相应PlanSection的问题。

+++ src/main/java/com/chris/gym/Plan/Dao/PlanDao.java
@@ -32,10 +32,11 @@ public class PlanDao {
     public Plan updatePlan(User user,Plan requestPlan){
         // 更新Plan的信息。
 
-        // 1. 检查Plan是否存在，存在则update，不存在则insert。
+        // 1. 获取Plan。
+        // 检查Plan是否存在，存在则update，不存在则insert。
         Plan responsePlan = null;
 
-        // 1.1 id不等于0，则数据库中可能存在。
+        // 1.1 id不等于0，则数据库中存在。
         if (requestPlan.getId() != 0){
             // 1.1.1 先尝试获取服务器上的Plan内容。
             try {
@@ -50,21 +51,22 @@ public class PlanDao {
             } catch (Exception e){
                 System.out.println("Error: 检查Plan是否存在中," + e + "." );
                 return null;
-//                return new ResponseEntity<String>("Error", HttpStatus.INTERNAL_SERVER_ERROR);
             }
         }
 
+        // 2. 修改或创建Plan.
         if( responsePlan == null ){
-            // 1.2 创建plan。
+            // 2.1 创建plan。
             try {
                 planMapper.insertPlan(requestPlan.getName(),user.getId(), requestPlan.getSeq());
                 responsePlan = planMapper.findPlanWithUserIdAndSeq(user.getId(), requestPlan.getSeq());
             } catch (Exception e){
                 System.out.println(e);
+                return null;
             }
 
         } else {
-            // 1.2 更新plan。
+            // 2.2 更新plan。
             if(responsePlan.getName().contentEquals(requestPlan.getName()) == false){
                 // 需要修改的plan和数据库中的plan名称不一致。
                 // 故update修改，否则不作任何改动。
@@ -76,55 +78,73 @@ public class PlanDao {
             }
         }
 
-        // 2. 下一步骤，对PlanSection和PlanRow进行检查修改。
-        //      如果数据库已存在这个Plan的某个Section，则直接修改参数。
+        // 3. 修改或者创建PlanSection。
+        //    如果数据库已存在这个Plan的某个Section，则直接修改参数。
         for(int i = 0; i<requestPlan.getSectionList().size(); i++){
-            if ( i < responsePlan.getSectionList().size()){
-                // 2.1 如果数据库已存在这个Plan的某个Section，
-                //       则直接传入对应的Section修改参数。
-                updateSection(
-                        responsePlan,
-                        responsePlan.getSectionList().get(i),
-                        requestPlan.getSectionList().get(i));
-            } else {
-                // 2.2 如果不存在，直接传入null，自动创建。
-                updateSection(
-                        responsePlan,
-                        null,
-                        requestPlan.getSectionList().get(i));
+            if (responsePlan.getSectionList() == null ){
+                responsePlan.setSectionList(new ArrayList<>());
             }
+
+            if ( i < responsePlan.getSectionList().size()){
+                    // 3.1 PlanSection存在。
+                    //    更新。
+                    PlanSection section = updateSection(
+                            responsePlan,
+                            responsePlan.getSectionList().get(i),
+                            requestPlan.getSectionList().get(i));
+                    responsePlan.getSectionList().set(i,section);
+                } else {
+                    // 3.2 不存在，直接传入null，创建一个。
+                    PlanSection section = updateSection(
+                            responsePlan,
+                            null,
+                            requestPlan.getSectionList().get(i));
+                    responsePlan.getSectionList().add(section);
+                }
 
         }
 
         // 3. 删除多余的section。
-        int difference = responsePlan.getSectionList().size() - requestPlan.getSectionList().size();
-        if(difference > 0){
-            // 从后往前删除。
-            // 比如目标数量是2，当前数量是4。
-            // i从2-3（3为4-1）。
-            for (int i = requestPlan.getSectionList().size() ;
-                 i < responsePlan.getSectionList().size() ;
-                 i++ ){
-                try {
-                    planMapper.deletePlanSection(responsePlan.getSectionList().get(i).getId());
-                } catch (Exception e) {
-                    System.out.println(e);
+        if (responsePlan.getSectionList() != null){
+            int difference = responsePlan.getSectionList().size() - requestPlan.getSectionList().size();
+            if(difference > 0){
+                // 从后往前删除。
+                // 比如目标数量是2，当前数量是4。
+                // i从2-3（3为4-1）。
+                for (int i = requestPlan.getSectionList().size() ;
+                     i < responsePlan.getSectionList().size() ;
+                     i++ ){
+                    try {
+                        planMapper.deletePlanSection(responsePlan.getSectionList().get(i).getId());
+                    } catch (Exception e) {
+                        System.out.println(e);
+                    }
                 }
             }
         }
 
+
         return responsePlan;
     }
 
-    private void updateSection(Plan plan, PlanSection currentSection , PlanSection requestSection){
+    public void deletePlan(int plan_id){
+        // 1. 获取计划总数量数量
+
+        // 2. 删除指定的Plan。
+        planMapper.deletePlan(plan_id);
+
+        // 3. 调整删除的计划后面的计划的seq值。
+    }
+
+
+    private PlanSection updateSection(Plan plan, PlanSection currentSection , PlanSection requestSection){
+
         // 检查section本身是否需要改动。
-
-
         PlanSection responseSection = currentSection;
 
         if( responseSection == null ){
-            // 数据库中不存在。
-            // 创建planSection。
+            // 1. 获取PlanSection，不存在则创建
+            // 1.1 创建planSection。
             try {
                 planMapper.insertPlanSection(plan.getUser_id(), plan.getId(), requestSection.getSeq(), requestSection.getSport().getId());
 
@@ -138,11 +158,11 @@ public class PlanDao {
             }
 
         } else {
-            // 更新planSection。
-            // 对比需要修改的和已经存在的"运动信息Sport"是否一致，否则update。
+            // 1.2 更新planSection。
+            // 1.2.1 对比需要修改的和已经存在的"运动信息Sport"是否一致，否则update。
             if(responseSection.getSport().getName().contentEquals(requestSection.getSport().getName()) == false){
                 try {
-                    // 检查sport_id。
+                    // 1.2.2 检查sport_id。
                     planMapper.updatePlanSectionSportId(responseSection.getId(), requestSection.getSport().getId());
                 } catch (Exception e){
                     System.out.println(e);
@@ -150,40 +170,52 @@ public class PlanDao {
             }
         }
 
-        // 处理更新Row。
+        // 2. 处理更新Row。
         for(int i=0; i<requestSection.getRowList().size(); i++){
-            if(i < responseSection.getRowList().size()) {
-                updateRow(
-                        responseSection,
-                        responseSection.getRowList().get(i),
-                        requestSection.getRowList().get(i));
-            } else {
-                // 当前数据库中不存在。
-                updateRow(
-                        responseSection,
-                        null,
-                        requestSection.getRowList().get(i));
-            }
-        }
-
-        // 删除多余的row。
-        int difference = responseSection.getRowList().size() - requestSection.getRowList().size();
-        if(difference > 0){
-            for (int i = requestSection.getRowList().size() ;
-                 i < responseSection.getRowList().size() ;
-                 i++){
-                try {
-                    planMapper.deletePlanRow(responseSection.getRowList().get(i).getId());
-                } catch (Exception e) {
-
-                    System.out.println(e);
+            if (responseSection.getRowList() != null) {
+                if(i < responseSection.getRowList().size()) {
+                    // 2.1 修改PlanRow.
+                    PlanRow row = updateRow(
+                            responseSection,
+                            responseSection.getRowList().get(i),
+                            requestSection.getRowList().get(i));
+                    responseSection.getRowList().set(i,row);
+//                    break;
+                } else {
+                    // 2.2 创建PlanRow
+                    // 当前数据库中不存在。
+                    PlanRow row = updateRow(
+                            responseSection,
+                            null,
+                            requestSection.getRowList().get(i));
+                    responseSection.getRowList().add(row);
                 }
             }
         }
 
+        // 3. 删除多余的row。
+        if (responseSection.getRowList() != null){
+            int difference = responseSection.getRowList().size() - requestSection.getRowList().size();
+            if(difference > 0){
+                for (int i = requestSection.getRowList().size() ;
+                     i < responseSection.getRowList().size() ;
+                     i++){
+                    try {
+                        planMapper.deletePlanRow(responseSection.getRowList().get(i).getId());
+                    } catch (Exception e) {
+
+                        System.out.println(e);
+                    }
+                }
+            }
+        }
+
+        return responseSection;
+
+
     }
 
-    private void updateRow(PlanSection responseSection, PlanRow currentRow, PlanRow requestRow){
+    private PlanRow updateRow(PlanSection responseSection, PlanRow currentRow, PlanRow requestRow){
 
         // 查看数据库中是否存在。
         PlanRow responseRow = currentRow;
@@ -231,6 +263,8 @@ public class PlanDao {
             }
 
         }
+
+        return responseRow;
     }
 
     private Plan parseToPlan(List<PlanSQLModel> temp) {
